+++ OpenRA.Mods.Common/Traits/Crushable.cs
@@ -9,10 +9,9 @@
 #endregion
 
 using System.Linq;
-using OpenRA.Mods.Common.Traits;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA.Traits
+namespace OpenRA.Mods.Common.Traits
 {
 	[Desc("This actor is crushable.")]
 	class CrushableInfo : ITraitInfo
